requirements.txt
# Core Dependencies
streamlit==1.29.0
pandas==2.1.3
numpy==1.26.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0

# Excel Processing
openpyxl==3.1.2
xlrd==2.0.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1  # Optional, but makes fuzzywuzzy faster

# AI/ML
scikit-learn==1.3.2
transformers==4.35.2
torch==2.2.0
sentence-transformers==2.2.2

# OCR
easyocr==1.7.1

# NLP
spacy==3.7.2
nltk==3.8.1

# Data Processing
regex==2023.10.3

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dotenv==1.0.0
jsonschema==4.20.0

# RAG and Vector Database
faiss-cpu==1.7.4
langchain==0.1.0
langchain-community==0.0.10
tiktoken==0.5.2